+++ packages/solverpy/src/solverpy/solver/smt/__init__.py
@@ -13,6 +13,7 @@ class solverpy.solver.smt.z3.Z3 extends solverpy.solver.shellsolver.ShellSolver
 class solverpy.solver.smt.bitwuzla.Bitwuzla extends solverpy.solver.shellsolver.ShellSolver
 class solverpy.solver.smt.llm2smt.Llm2smt extends solverpy.solver.shellsolver.ShellSolver
 class solverpy.solver.smt.primo.Primo extends solverpy.solver.shellsolver.ShellSolver
+class solverpy.solver.smt.opensmt.Opensmt extends solverpy.solver.shellsolver.ShellSolver
 
 ```
 
@@ -23,13 +24,15 @@ class solverpy.solver.smt.primo.Primo extends solverpy.solver.shellsolver.ShellS
 | [`Bitwuzla`][solverpy.solver.smt.bitwuzla.Bitwuzla] | `bitwuzla` | Specialised for bit-vectors and floating point |
 | [`Llm2smt`][solverpy.solver.smt.llm2smt.Llm2smt] | `llm2smt` | QF_EUF solver |
 | [`Primo`][solverpy.solver.smt.primo.Primo] | `primo` | QF_LRA and QF_UF solver |
+| [`Opensmt`][solverpy.solver.smt.opensmt.Opensmt] | `opensmt` | SMT solver with interpolation and proofs |
 
 """
 
 from .bitwuzla import Bitwuzla
 from .cvc5 import Cvc5
 from .llm2smt import Llm2smt
+from .opensmt import Opensmt
 from .primo import Primo
 from .z3 import Z3
 
-__all__ = ["Bitwuzla", "Cvc5", "Llm2smt", "Primo", "Z3"]
+__all__ = ["Bitwuzla", "Cvc5", "Llm2smt", "Opensmt", "Primo", "Z3"]